+++ tests/integration/test_litestar_session_backend.py
@@ -4,13 +4,16 @@ import asyncio
 import datetime
 import uuid
 from collections.abc import AsyncGenerator, Generator
-from typing import Dict, Optional, Union
+from functools import partial
+from typing import Optional, Union
+from unittest.mock import Mock
 
 import pytest
 from litestar import Litestar, Request, get, post
 from litestar.middleware.session import SessionMiddleware
 from litestar.middleware.session.server_side import ServerSideSessionConfig
-from litestar.testing import AsyncTestClient, create_test_client
+from litestar.stores.base import Store
+from litestar.testing import AsyncTestClient
 from sqlalchemy import Engine, create_engine, select
 from sqlalchemy.ext.asyncio import AsyncEngine, AsyncSession, async_sessionmaker, create_async_engine
 from sqlalchemy.orm import Session, sessionmaker
@@ -25,10 +28,16 @@ from advanced_alchemy.extensions.litestar.session import (
 )
 
 
-class SessionModel(SessionModelMixin, UUIDv7Base):
-    """Test session model."""
+class AsyncSessionModel(SessionModelMixin, UUIDv7Base):
+    """Test session model for async tests."""
 
-    __tablename__ = "test_sessions"
+    __tablename__ = "async_test_sessions"
+
+
+class SyncSessionModel(SessionModelMixin, UUIDv7Base):
+    """Test session model for sync tests."""
+
+    __tablename__ = "sync_test_sessions"
 
 
 @pytest.fixture()
@@ -44,10 +53,22 @@ async def async_engine() -> AsyncGenerator[AsyncEngine, None]:
 @pytest.fixture()
 def sync_engine() -> Generator[Engine, None, None]:
     """Create a sync SQLite engine for testing."""
-    engine = create_engine("sqlite:///:memory:")
-    UUIDv7Base.metadata.create_all(engine)
-    yield engine
-    engine.dispose()
+    import os
+    import tempfile
+
+    # Create a temporary file for the database
+    fd, db_path = tempfile.mkstemp(suffix=".db")
+    os.close(fd)
+
+    try:
+        engine = create_engine(f"sqlite:///{db_path}")
+        UUIDv7Base.metadata.create_all(engine)
+        yield engine
+        engine.dispose()
+    finally:
+        # Clean up the temporary file
+        if os.path.exists(db_path):
+            os.unlink(db_path)
 
 
 @pytest.fixture()
@@ -62,19 +83,22 @@ def sync_session_factory(sync_engine: Engine) -> sessionmaker[Session]:
     return sessionmaker(sync_engine, expire_on_commit=False)
 
 
+@pytest.fixture()
+def mock_store() -> Store:
+    """Create a mock store for testing."""
+    return Mock(spec=Store)
+
+
 class TestAsyncSessionBackendIntegration:
     """Integration tests for async session backend."""
 
     @pytest.fixture()
-    async def async_config(self, async_session_factory: async_sessionmaker[AsyncSession]) -> SQLAlchemyAsyncConfig:
-        """Create async config with test session factory."""
-        config = SQLAlchemyAsyncConfig(
-            connection_string="sqlite+aiosqlite:///:memory:",
+    async def async_config(self, async_engine: AsyncEngine) -> SQLAlchemyAsyncConfig:
+        """Create async config with test engine."""
+        return SQLAlchemyAsyncConfig(
+            engine_instance=async_engine,
             session_dependency_key="db_session",
         )
-        # Override the session factory
-        config.get_session = async_session_factory
-        return config
 
     @pytest.fixture()
     async def async_backend(self, async_config: SQLAlchemyAsyncConfig) -> SQLAlchemyAsyncSessionBackend:
@@ -82,94 +106,99 @@ class TestAsyncSessionBackendIntegration:
         return SQLAlchemyAsyncSessionBackend(
             config=ServerSideSessionConfig(max_age=3600),
             alchemy_config=async_config,
-            model=SessionModel,
+            model=AsyncSessionModel,
         )
 
     async def test_session_lifecycle(
         self,
         async_backend: SQLAlchemyAsyncSessionBackend,
         async_session_factory: async_sessionmaker[AsyncSession],
+        mock_store: Store,
     ) -> None:
         """Test complete session lifecycle: create, retrieve, update, delete."""
         session_id = str(uuid.uuid4())
         original_data = b"test_data_123"
         updated_data = b"updated_data_456"
-        store = None  # Mock store
 
         # Create session
-        await async_backend.set(session_id, original_data, store)
+        await async_backend.set(session_id, original_data, mock_store)
 
         # Verify in database
         async with async_session_factory() as db_session:
-            result = await db_session.execute(select(SessionModel).where(SessionModel.session_id == session_id))
+            result = await db_session.execute(
+                select(AsyncSessionModel).where(AsyncSessionModel.session_id == session_id)
+            )
             session_obj = result.scalar_one()
-            assert session_obj.data == original_data
-            assert not session_obj.is_expired
+            data_in_db = session_obj.data
+            is_expired = session_obj.is_expired
+            assert data_in_db == original_data
+            assert not is_expired
 
         # Retrieve session
-        retrieved_data = await async_backend.get(session_id, store)
+        retrieved_data = await async_backend.get(session_id, mock_store)
         assert retrieved_data == original_data
 
         # Update session
-        await async_backend.set(session_id, updated_data, store)
+        await async_backend.set(session_id, updated_data, mock_store)
 
         # Verify update
-        retrieved_data = await async_backend.get(session_id, store)
+        retrieved_data = await async_backend.get(session_id, mock_store)
         assert retrieved_data == updated_data
 
         # Delete session
-        await async_backend.delete(session_id, store)
+        await async_backend.delete(session_id, mock_store)
 
         # Verify deletion
-        retrieved_data = await async_backend.get(session_id, store)
+        retrieved_data = await async_backend.get(session_id, mock_store)
         assert retrieved_data is None
 
     async def test_session_expiration(
         self,
         async_session_factory: async_sessionmaker[AsyncSession],
+        mock_store: Store,
     ) -> None:
         """Test session expiration handling."""
         # Create backend with very short expiration
         config = SQLAlchemyAsyncConfig(
-            connection_string="sqlite+aiosqlite:///:memory:",
+            engine_instance=async_session_factory.kw["bind"],
             session_dependency_key="db_session",
         )
-        config.get_session = async_session_factory
 
         backend = SQLAlchemyAsyncSessionBackend(
             config=ServerSideSessionConfig(max_age=1),  # 1 second
             alchemy_config=config,
-            model=SessionModel,
+            model=AsyncSessionModel,
         )
 
         session_id = str(uuid.uuid4())
         data = b"expires_soon"
-        store = None
 
         # Create session
-        await backend.set(session_id, data, store)
+        await backend.set(session_id, data, mock_store)
 
         # Verify it exists
-        assert await backend.get(session_id, store) == data
+        assert await backend.get(session_id, mock_store) == data
 
         # Wait for expiration
         await asyncio.sleep(2)
 
         # Should return None and delete expired session
-        assert await backend.get(session_id, store) is None
+        assert await backend.get(session_id, mock_store) is None
 
         # Verify it's deleted from database
         async with async_session_factory() as db_session:
-            result = await db_session.execute(select(SessionModel).where(SessionModel.session_id == session_id))
+            result = await db_session.execute(
+                select(AsyncSessionModel).where(AsyncSessionModel.session_id == session_id)
+            )
             assert result.scalar_one_or_none() is None
 
     async def test_delete_expired_sessions(
         self,
         async_backend: SQLAlchemyAsyncSessionBackend,
         async_session_factory: async_sessionmaker[AsyncSession],
+        mock_store: Store,
     ) -> None:
         """Test bulk deletion of expired sessions."""
-        store = None
         now = datetime.datetime.now(datetime.timezone.utc)
 
         # Create mix of expired and active sessions
@@ -179,7 +208,7 @@ class TestAsyncSessionBackendIntegration:
         # Insert expired sessions directly
         async with async_session_factory() as db_session:
             for sid in expired_ids:
-                session = SessionModel(
+                session = AsyncSessionModel(
                     session_id=sid,
                     data=b"expired",
                     expires_at=now - datetime.timedelta(hours=1),
@@ -189,14 +218,14 @@ class TestAsyncSessionBackendIntegration:
 
         # Create active sessions through backend
         for sid in active_ids:
-            await async_backend.set(sid, b"active", store)
+            await async_backend.set(sid, b"active", mock_store)
 
         # Delete expired
         await async_backend.delete_expired()
 
         # Verify only active sessions remain
         async with async_session_factory() as db_session:
-            result = await db_session.execute(select(SessionModel.session_id))
+            result = await db_session.execute(select(AsyncSessionModel.session_id))
             remaining_ids = {row[0] for row in result}
             assert remaining_ids == set(active_ids)
 
@@ -204,49 +233,48 @@ class TestAsyncSessionBackendIntegration:
         self,
         async_backend: SQLAlchemyAsyncSessionBackend,
         async_session_factory: async_sessionmaker[AsyncSession],
+        mock_store: Store,
     ) -> None:
         """Test deletion of all sessions."""
-        store = None
-
         # Create multiple sessions
         session_ids = [str(uuid.uuid4()) for _ in range(5)]
         for sid in session_ids:
-            await async_backend.set(sid, b"data", store)
+            await async_backend.set(sid, b"data", mock_store)
 
         # Verify they exist
         async with async_session_factory() as db_session:
             from sqlalchemy import func
 
-            count = await db_session.scalar(select(func.count()).select_from(SessionModel))
+            count = await db_session.scalar(select(func.count()).select_from(AsyncSessionModel))
             assert count == 5
 
         # Delete all
-        await async_backend.delete_all(store)
+        await async_backend.delete_all(mock_store)
 
         # Verify all deleted
         async with async_session_factory() as db_session:
             from sqlalchemy import func
 
-            count = await db_session.scalar(select(func.count()).select_from(SessionModel))
+            count = await db_session.scalar(select(func.count()).select_from(AsyncSessionModel))
             assert count == 0
 
     async def test_concurrent_session_access(
         self,
         async_backend: SQLAlchemyAsyncSessionBackend,
+        mock_store: Store,
     ) -> None:
         """Test concurrent access to sessions."""
-        store = None
         session_id = str(uuid.uuid4())
 
         # Initial set
-        await async_backend.set(session_id, b"initial", store)
+        await async_backend.set(session_id, b"initial", mock_store)
 
         # Concurrent reads and writes
         async def read_session(n: int) -> Optional[bytes]:
-            return await async_backend.get(session_id, store)
+            return await async_backend.get(session_id, mock_store)
 
         async def write_session(n: int) -> None:
-            await async_backend.set(session_id, f"data_{n}".encode(), store)
+            await async_backend.set(session_id, f"data_{n}".encode(), mock_store)
 
         # Run concurrent operations
         tasks = []
@@ -262,7 +290,7 @@ class TestAsyncSessionBackendIntegration:
                 raise result
 
         # Verify session still exists and has valid data
-        final_data = await async_backend.get(session_id, store)
+        final_data = await async_backend.get(session_id, mock_store)
         assert final_data is not None
         assert final_data.startswith(b"data_")
 
@@ -271,15 +299,12 @@ class TestSyncSessionBackendIntegration:
     """Integration tests for sync session backend."""
 
     @pytest.fixture()
-    def sync_config(self, sync_session_factory: sessionmaker[Session]) -> SQLAlchemySyncConfig:
-        """Create sync config with test session factory."""
-        config = SQLAlchemySyncConfig(
-            connection_string="sqlite:///:memory:",
+    def sync_config(self, sync_engine: Engine) -> SQLAlchemySyncConfig:
+        """Create sync config with test engine."""
+        return SQLAlchemySyncConfig(
+            engine_instance=sync_engine,
             session_dependency_key="db_session",
         )
-        # Override the session factory
-        config.get_session = sync_session_factory
-        return config
 
     @pytest.fixture()
     def sync_backend(self, sync_config: SQLAlchemySyncConfig) -> SQLAlchemySyncSessionBackend:
@@ -287,7 +312,7 @@ class TestSyncSessionBackendIntegration:
         return SQLAlchemySyncSessionBackend(
             config=ServerSideSessionConfig(max_age=3600),
             alchemy_config=sync_config,
-            model=SessionModel,
+            model=SyncSessionModel,
         )
 
     @pytest.mark.asyncio()
@@ -295,39 +320,41 @@ class TestSyncSessionBackendIntegration:
         self,
         sync_backend: SQLAlchemySyncSessionBackend,
         sync_session_factory: sessionmaker[Session],
+        mock_store: Store,
     ) -> None:
         """Test complete session lifecycle with sync backend."""
         session_id = str(uuid.uuid4())
         original_data = b"sync_test_data"
         updated_data = b"sync_updated_data"
-        store = None
 
         # Create session
-        await sync_backend.set(session_id, original_data, store)
+        await sync_backend.set(session_id, original_data, mock_store)
 
         # Verify in database
         with sync_session_factory() as db_session:
-            result = db_session.execute(select(SessionModel).where(SessionModel.session_id == session_id))
+            result = db_session.execute(select(SyncSessionModel).where(SyncSessionModel.session_id == session_id))
             session_obj = result.scalar_one()
-            assert session_obj.data == original_data
-            assert not session_obj.is_expired
+            data_in_db = session_obj.data
+            is_expired = session_obj.is_expired
+            assert data_in_db == original_data
+            assert not is_expired
 
         # Retrieve session
-        retrieved_data = await sync_backend.get(session_id, store)
+        retrieved_data = await sync_backend.get(session_id, mock_store)
         assert retrieved_data == original_data
 
         # Update session
-        await sync_backend.set(session_id, updated_data, store)
+        await sync_backend.set(session_id, updated_data, mock_store)
 
         # Verify update
-        retrieved_data = await sync_backend.get(session_id, store)
+        retrieved_data = await sync_backend.get(session_id, mock_store)
         assert retrieved_data == updated_data
 
         # Delete session
-        await sync_backend.delete(session_id, store)
+        await sync_backend.delete(session_id, mock_store)
 
         # Verify deletion
-        retrieved_data = await sync_backend.get(session_id, store)
+        retrieved_data = await sync_backend.get(session_id, mock_store)
         assert retrieved_data is None
 
     @pytest.mark.asyncio()
@@ -335,9 +362,9 @@ class TestSyncSessionBackendIntegration:
         self,
         sync_backend: SQLAlchemySyncSessionBackend,
         sync_session_factory: sessionmaker[Session],
+        mock_store: Store,
     ) -> None:
         """Test bulk deletion of expired sessions with sync backend."""
-        store = None
         now = datetime.datetime.now(datetime.timezone.utc)
 
         # Create mix of expired and active sessions
@@ -347,7 +374,7 @@ class TestSyncSessionBackendIntegration:
         # Insert expired sessions directly
         with sync_session_factory() as db_session:
             for sid in expired_ids:
-                session = SessionModel(
+                session = SyncSessionModel(
                     session_id=sid,
                     data=b"expired",
                     expires_at=now - datetime.timedelta(hours=1),
@@ -357,14 +384,14 @@ class TestSyncSessionBackendIntegration:
 
         # Create active sessions through backend
         for sid in active_ids:
-            await sync_backend.set(sid, b"active", store)
+            await sync_backend.set(sid, b"active", mock_store)
 
         # Delete expired
         await sync_backend.delete_expired()
 
         # Verify only active sessions remain
         with sync_session_factory() as db_session:
-            result = db_session.execute(select(SessionModel.session_id))
+            result = db_session.execute(select(SyncSessionModel.session_id))
             remaining_ids = {row[0] for row in result}
             assert remaining_ids == set(active_ids)
 
@@ -378,38 +405,37 @@ class TestLitestarIntegration:
     ) -> None:
         """Test async session backend with Litestar middleware."""
         config = SQLAlchemyAsyncConfig(
-            connection_string="sqlite+aiosqlite:///:memory:",
+            engine_instance=async_session_factory.kw["bind"],
             session_dependency_key="db_session",
         )
-        config.get_session = async_session_factory
 
         backend = SQLAlchemyAsyncSessionBackend(
             config=ServerSideSessionConfig(max_age=3600, key="test-session"),
             alchemy_config=config,
-            model=SessionModel,
+            model=AsyncSessionModel,
         )
 
         @get("/set")
-        async def set_session(request: Request) -> Dict[str, str]:
+        async def set_session(request: Request) -> dict[str, str]:
             request.session["user_id"] = "123"
             request.session["username"] = "testuser"
             return {"status": "session set"}
 
         @get("/get")
-        async def get_session(request: Request) -> Dict[str, Optional[str]]:
+        async def get_session(request: Request) -> dict[str, Optional[str]]:
             return {
                 "user_id": request.session.get("user_id"),
                 "username": request.session.get("username"),
             }
 
         @post("/clear")
-        async def clear_session(request: Request) -> Dict[str, str]:
+        async def clear_session(request: Request) -> dict[str, str]:
             request.clear_session()
             return {"status": "session cleared"}
 
         app = Litestar(
             route_handlers=[set_session, get_session, clear_session],
-            middleware=[SessionMiddleware(backend=backend, key="test-session")],
+            middleware=[partial(SessionMiddleware, backend=backend)],
         )
 
         async with AsyncTestClient(app=app) as client:
@@ -425,7 +451,7 @@ class TestLitestarIntegration:
 
             # Clear session
             response = await client.post("/clear")
-            assert response.status_code == 200
+            assert response.status_code == 201
             assert response.json() == {"status": "session cleared"}
 
             # Verify cleared
@@ -433,50 +459,49 @@ class TestLitestarIntegration:
             assert response.status_code == 200
             assert response.json() == {"user_id": None, "username": None}
 
-    def test_sync_session_middleware(
+    async def test_sync_session_middleware(
         self,
         sync_session_factory: sessionmaker[Session],
     ) -> None:
         """Test sync session backend with Litestar middleware."""
         config = SQLAlchemySyncConfig(
-            connection_string="sqlite:///:memory:",
+            engine_instance=sync_session_factory.kw["bind"],
             session_dependency_key="db_session",
         )
-        config.get_session = sync_session_factory
 
         backend = SQLAlchemySyncSessionBackend(
             config=ServerSideSessionConfig(max_age=3600, key="test-session"),
             alchemy_config=config,
-            model=SessionModel,
+            model=SyncSessionModel,
         )
 
         @get("/set")
-        def set_session(request: Request) -> Dict[str, str]:
+        def set_session(request: Request) -> dict[str, str]:
             request.session["counter"] = request.session.get("counter", 0) + 1
             return {"counter": request.session["counter"]}
 
         @get("/get")
-        def get_session(request: Request) -> Dict[str, Optional[int]]:
+        def get_session(request: Request) -> dict[str, Optional[int]]:
             return {"counter": request.session.get("counter")}
 
         app = Litestar(
             route_handlers=[set_session, get_session],
-            middleware=[SessionMiddleware(backend=backend, key="test-session")],
+            middleware=[partial(SessionMiddleware, backend=backend)],
         )
 
-        with create_test_client(app=app) as client:
+        async with AsyncTestClient(app=app) as client:
             # Initial set
-            response = client.get("/set")
+            response = await client.get("/set")
             assert response.status_code == 200
             assert response.json() == {"counter": 1}
 
             # Increment counter
-            response = client.get("/set")
+            response = await client.get("/set")
             assert response.status_code == 200
             assert response.json() == {"counter": 2}
 
             # Get current value
-            response = client.get("/get")
+            response = await client.get("/get")
             assert response.status_code == 200
             assert response.json() == {"counter": 2}
 
@@ -486,26 +511,25 @@ class TestLitestarIntegration:
     ) -> None:
         """Test that sessions persist across multiple requests."""
         config = SQLAlchemyAsyncConfig(
-            connection_string="sqlite+aiosqlite:///:memory:",
+            engine_instance=async_session_factory.kw["bind"],
             session_dependency_key="db_session",
         )
-        config.get_session = async_session_factory
 
         backend = SQLAlchemyAsyncSessionBackend(
             config=ServerSideSessionConfig(max_age=3600, key="test-session"),
             alchemy_config=config,
-            model=SessionModel,
+            model=AsyncSessionModel,
         )
 
         @get("/login")
-        async def login(request: Request) -> Dict[str, str]:
+        async def login(request: Request) -> dict[str, str]:
             request.session["authenticated"] = True
             request.session["user_id"] = "user123"
             request.session["login_time"] = datetime.datetime.now(datetime.timezone.utc).isoformat()
             return {"status": "logged in"}
 
         @get("/profile")
-        async def profile(request: Request) -> Dict[str, Union[str, bool]]:
+        async def profile(request: Request) -> dict[str, Union[str, bool, None]]:
             if not request.session.get("authenticated"):
                 return {"error": "not authenticated"}
             return {
@@ -514,13 +538,13 @@ class TestLitestarIntegration:
             }
 
         @post("/logout")
-        async def logout(request: Request) -> Dict[str, str]:
+        async def logout(request: Request) -> dict[str, str]:
             request.clear_session()
             return {"status": "logged out"}
 
         app = Litestar(
             route_handlers=[login, profile, logout],
-            middleware=[SessionMiddleware(backend=backend, key="test-session")],
+            middleware=[partial(SessionMiddleware, backend=backend)],
         )
 
         async with AsyncTestClient(app=app) as client:
