+++ tests/unit/test_extensions/test_litestar/test_session.py
@@ -1,7 +1,8 @@
 import datetime
+import sys
 from collections.abc import Awaitable, Generator
 from typing import Any, Callable, TypeVar
-from unittest.mock import MagicMock, Mock, patch
+from unittest.mock import AsyncMock, MagicMock, Mock, patch
 
 import pytest
 from litestar.middleware.session.server_side import ServerSideSessionConfig
@@ -161,7 +162,7 @@ def async_backend(async_backend_config: SQLAlchemyAsyncSessionBackend) -> SQLAlc
 async def test_async_backend_get_session_obj_found(
     async_backend: SQLAlchemyAsyncSessionBackend,
     mock_session_model: type[SessionModelMixin],
-    mock_async_session: MagicMock,  # Inject mock session directly
+    mock_async_session: MagicMock,
 ) -> None:
     """Test _get_session_obj finds an existing session."""
     mock_scalar_result = MagicMock()
@@ -173,7 +174,6 @@ async def test_async_backend_get_session_obj_found(
 
     assert session_obj is expected_session
     mock_async_session.scalars.assert_called_once()
-    # Check the select statement structure loosely
     call_args, _ = mock_async_session.scalars.call_args
     select_stmt = call_args[0]
     assert str(select_stmt).startswith(f"SELECT {mock_session_model.__tablename__}")
@@ -182,7 +182,7 @@ async def test_async_backend_get_session_obj_found(
 
 async def test_async_backend_get_session_obj_not_found(
     async_backend: SQLAlchemyAsyncSessionBackend,
-    mock_async_session: MagicMock,  # Inject mock session
+    mock_async_session: MagicMock,
 ) -> None:
     """Test _get_session_obj returns None when session not found."""
     mock_scalar_result = MagicMock()
@@ -202,7 +202,7 @@ async def test_async_backend_get_existing_not_expired(
     async_backend: SQLAlchemyAsyncSessionBackend,
     mock_async_config: MagicMock,
     mock_session_model: type[SessionModelMixin],
-    mock_async_session: MagicMock,  # Inject mock session
+    mock_async_session: MagicMock,
 ) -> None:
     """Test getting an existing, non-expired session."""
     now = datetime.datetime.now(datetime.timezone.utc)
@@ -219,7 +219,7 @@ async def test_async_backend_get_existing_not_expired(
         assert result == session_data
         mock_get_obj.assert_awaited_once_with(db_session=mock_async_session, session_id=session_id)
         # Check expiry was updated
-        assert mock_session_obj.expires_at > now  # Compare with current time instead of old expiry
+        assert mock_session_obj.expires_at > now
         mock_async_session.commit.assert_awaited_once()
         mock_async_session.delete.assert_not_called()
 
@@ -231,7 +231,7 @@ async def test_async_backend_get_existing_expired(
     async_backend: SQLAlchemyAsyncSessionBackend,
     mock_async_config: MagicMock,
     mock_session_model: type[SessionModelMixin],
-    mock_async_session: MagicMock,  # Inject mock session
+    mock_async_session: MagicMock,
 ) -> None:
     """Test getting an expired session returns None and deletes it."""
     now = datetime.datetime.now(datetime.timezone.utc)
@@ -254,7 +254,7 @@ async def test_async_backend_get_existing_expired(
 async def test_async_backend_get_non_existent(
     async_backend: SQLAlchemyAsyncSessionBackend,
     mock_async_config: MagicMock,
-    mock_async_session: MagicMock,  # Inject mock session
+    mock_async_session: MagicMock,
 ) -> None:
     """Test getting a non-existent session returns None."""
     session_id = "non_existent_session"
@@ -274,7 +274,7 @@ async def test_async_backend_set_new_session(
     async_backend: SQLAlchemyAsyncSessionBackend,
     mock_async_config: MagicMock,
     mock_session_model: type[SessionModelMixin],
-    mock_async_session: MagicMock,  # Inject mock session
+    mock_async_session: MagicMock,
 ) -> None:
     """Test setting a new session."""
     now = datetime.datetime.now(datetime.timezone.utc)
@@ -302,7 +302,7 @@ async def test_async_backend_set_update_existing_session(
     async_backend: SQLAlchemyAsyncSessionBackend,
     mock_async_config: MagicMock,
     mock_session_model: type[SessionModelMixin],
-    mock_async_session: MagicMock,  # Inject mock session
+    mock_async_session: MagicMock,
 ) -> None:
     """Test updating an existing session's data and expiry."""
     now = datetime.datetime.now(datetime.timezone.utc)
@@ -329,14 +329,13 @@ async def test_async_backend_delete_existing(
     async_backend: SQLAlchemyAsyncSessionBackend,
     mock_async_config: MagicMock,
     mock_session_model: type[SessionModelMixin],
-    mock_async_session: MagicMock,  # Inject mock session
+    mock_async_session: MagicMock,
 ) -> None:
     """Test deleting an existing session."""
     session_id = "delete_me"
     await async_backend.delete(session_id, store=Mock())
 
     mock_async_session.execute.assert_awaited_once()
-    # Check the delete statement structure loosely
     call_args, _ = mock_async_session.execute.call_args
     delete_stmt = call_args[0]
     assert str(delete_stmt).startswith(f"DELETE FROM {mock_session_model.__tablename__}")
@@ -348,7 +347,7 @@ async def test_async_backend_delete_existing(
 async def test_async_backend_delete_non_existent(
     async_backend: SQLAlchemyAsyncSessionBackend,
     mock_async_config: MagicMock,
-    mock_async_session: MagicMock,  # Inject mock session
+    mock_async_session: MagicMock,
 ) -> None:
     """Test deleting a non-existent session (fails silently)."""
     session_id = "i_dont_exist"
@@ -363,16 +362,15 @@ async def test_async_backend_delete_all(
     async_backend: SQLAlchemyAsyncSessionBackend,
     mock_async_config: MagicMock,
     mock_session_model: type[SessionModelMixin],
-    mock_async_session: MagicMock,  # Inject mock session
+    mock_async_session: MagicMock,
 ) -> None:
     """Test deleting all sessions."""
     await async_backend.delete_all(store=Mock())
 
     mock_async_session.execute.assert_awaited_once()
-    # Check the delete statement structure loosely
     call_args, _ = mock_async_session.execute.call_args
     delete_stmt = call_args[0]
-    assert str(delete_stmt) == f"DELETE FROM {mock_session_model.__tablename__}"  # No WHERE clause
+    assert str(delete_stmt) == f"DELETE FROM {mock_session_model.__tablename__}"
     mock_async_session.commit.assert_awaited_once()
 
 
@@ -382,7 +380,7 @@ async def test_async_backend_delete_expired(
     mock_async_config: MagicMock,
     mock_session_model: type[SessionModelMixin],
     monkeypatch: MonkeyPatch,
-    mock_async_session: MagicMock,  # Inject mock session
+    mock_async_session: MagicMock,
 ) -> None:
     """Test deleting expired sessions."""
     mock_async_session = mock_async_config.get_session.return_value
@@ -396,12 +394,10 @@ async def test_async_backend_delete_expired(
     await async_backend.delete_expired()
 
     mock_async_session.execute.assert_awaited_once()
-    # Check the delete statement structure loosely
     call_args, _ = mock_async_session.execute.call_args
     delete_stmt = call_args[0]
     assert str(delete_stmt).startswith(f"DELETE FROM {mock_session_model.__tablename__}")
-    # Check for the expiration condition in the WHERE clause
-    assert "WHERE" in str(delete_stmt) and "now() >" in str(delete_stmt)  # Match actual SQL
+    assert "WHERE" in str(delete_stmt) and "now() >" in str(delete_stmt)
     mock_async_session.commit.assert_awaited_once()
 
 
@@ -436,7 +432,7 @@ def sync_backend(mock_sync_config: MagicMock) -> Generator[SQLAlchemySyncSession
 async def test_sync_backend_get_wraps_sync_call(
     sync_backend: SQLAlchemySyncSessionBackend,
     mock_sync_config: MagicMock,
-    mock_sync_session: MagicMock,  # Inject mock session
+    mock_sync_session: MagicMock,
 ) -> None:
     """Test that sync backend's async get() calls the internal _get_sync."""
     session_id = "sync_test_get"
@@ -456,7 +452,7 @@ async def test_sync_backend_get_wraps_sync_call(
 async def test_sync_backend_set_wraps_sync_call(
     sync_backend: SQLAlchemySyncSessionBackend,
     mock_sync_config: MagicMock,
-    mock_sync_session: MagicMock,  # Inject mock session
+    mock_sync_session: MagicMock,
 ) -> None:
     """Test that sync backend's async set() calls the internal _set_sync."""
     session_id = "sync_test_set"
@@ -475,7 +471,7 @@ async def test_sync_backend_set_wraps_sync_call(
 async def test_sync_backend_delete_wraps_sync_call(
     sync_backend: SQLAlchemySyncSessionBackend,
     mock_sync_config: MagicMock,
-    mock_sync_session: MagicMock,  # Inject mock session
+    mock_sync_session: MagicMock,
 ) -> None:
     """Test that sync backend's async delete() calls the internal _delete_sync."""
     session_id = "sync_test_delete"
@@ -493,7 +489,7 @@ async def test_sync_backend_delete_wraps_sync_call(
 async def test_sync_backend_delete_all_wraps_sync_call(
     sync_backend: SQLAlchemySyncSessionBackend,
     mock_sync_config: MagicMock,
-    mock_sync_session: MagicMock,  # Inject mock session
+    mock_sync_session: MagicMock,
 ) -> None:
     """Test that sync backend's async delete_all() calls the internal _delete_all_sync."""
     mock_sync_session = mock_sync_config.get_session.return_value
@@ -510,7 +506,7 @@ async def test_sync_backend_delete_all_wraps_sync_call(
 async def test_sync_backend_delete_expired_wraps_sync_call(
     sync_backend: SQLAlchemySyncSessionBackend,
     mock_sync_config: MagicMock,
-    mock_sync_session: MagicMock,  # Inject mock session
+    mock_sync_session: MagicMock,
 ) -> None:
     """Test that sync backend's async delete_expired() calls the internal _delete_expired_sync."""
     mock_sync_session = mock_sync_config.get_session.return_value
@@ -532,7 +528,7 @@ def test_sync_backend_internal_get_sync_existing_not_expired(
     mock_sync_config: MagicMock,
     mock_session_model: type[SessionModelMixin],
     monkeypatch: MonkeyPatch,
-    mock_sync_session: MagicMock,  # Inject mock session
+    mock_sync_session: MagicMock,
 ) -> None:
     """Test the internal _get_sync method for existing, non-expired session."""
     now = datetime.datetime.now(datetime.timezone.utc)
@@ -554,7 +550,7 @@ def test_sync_backend_internal_get_sync_existing_not_expired(
 
         assert result == session_data
         mock_get_obj.assert_called_once_with(db_session=mock_sync_session, session_id=session_id)
-        assert mock_session_obj.expires_at > now  # Compare with current time instead of old expiry
+        assert mock_session_obj.expires_at > now
         mock_sync_session.commit.assert_called_once()
         mock_sync_session.delete.assert_not_called()
 
@@ -564,7 +560,7 @@ def test_sync_backend_internal_get_sync_existing_expired(
     mock_sync_config: MagicMock,
     mock_session_model: type[SessionModelMixin],
     monkeypatch: MonkeyPatch,
-    mock_sync_session: MagicMock,  # Inject mock session
+    mock_sync_session: MagicMock,
 ) -> None:
     """Test the internal _get_sync method for an expired session."""
     now = datetime.datetime.now(datetime.timezone.utc)
@@ -592,7 +588,7 @@ def test_sync_backend_internal_get_sync_existing_expired(
 def test_sync_backend_internal_get_sync_non_existent(
     sync_backend: SQLAlchemySyncSessionBackend,
     mock_sync_config: MagicMock,
-    mock_sync_session: MagicMock,  # Inject mock session
+    mock_sync_session: MagicMock,
 ) -> None:
     """Test the internal _get_sync method for a non-existent session."""
     session_id = "sync_non_existent"
@@ -614,7 +610,7 @@ def test_sync_backend_internal_set_sync_new(
     mock_sync_config: MagicMock,
     mock_session_model: type[SessionModelMixin],
     monkeypatch: MonkeyPatch,
-    mock_sync_session: MagicMock,  # Inject mock session
+    mock_sync_session: MagicMock,
 ) -> None:
     """Test the internal _set_sync method for a new session."""
     now = datetime.datetime.now(datetime.timezone.utc)
@@ -647,7 +643,7 @@ def test_sync_backend_internal_set_sync_update(
     mock_sync_config: MagicMock,
     mock_session_model: type[SessionModelMixin],
     monkeypatch: MonkeyPatch,
-    mock_sync_session: MagicMock,  # Inject mock session
+    mock_sync_session: MagicMock,
 ) -> None:
     """Test the internal _set_sync method updating an existing session."""
     now = datetime.datetime.now(datetime.timezone.utc)
@@ -679,7 +675,7 @@ def test_sync_backend_internal_delete_sync(
     sync_backend: SQLAlchemySyncSessionBackend,
     mock_sync_config: MagicMock,
     mock_session_model: type[SessionModelMixin],
-    mock_sync_session: MagicMock,  # Inject mock session
+    mock_sync_session: MagicMock,
 ) -> None:
     """Test the internal _delete_sync method."""
     session_id = "sync_delete_me"
@@ -691,7 +687,6 @@ def test_sync_backend_internal_delete_sync(
     sync_backend._delete_sync(session_id)  # pyright: ignore [reportPrivateUsage]
 
     mock_sync_session.execute.assert_called_once()
-    # Check the delete statement structure loosely
     call_args, _ = mock_sync_session.execute.call_args
     delete_stmt = call_args[0]
     assert str(delete_stmt).startswith(f"DELETE FROM {mock_session_model.__tablename__}")
@@ -703,7 +698,7 @@ def test_sync_backend_internal_delete_all_sync(
     sync_backend: SQLAlchemySyncSessionBackend,
     mock_sync_config: MagicMock,
     mock_session_model: type[SessionModelMixin],
-    mock_sync_session: MagicMock,  # Inject mock session
+    mock_sync_session: MagicMock,
 ) -> None:
     """Test the internal _delete_all_sync method."""
     mock_sync_session = MagicMock(spec=SyncSession)
@@ -714,10 +709,9 @@ def test_sync_backend_internal_delete_all_sync(
     sync_backend._delete_all_sync()  # pyright: ignore [reportPrivateUsage]
 
     mock_sync_session.execute.assert_called_once()
-    # Check the delete statement structure loosely
     call_args, _ = mock_sync_session.execute.call_args
     delete_stmt = call_args[0]
-    assert str(delete_stmt) == f"DELETE FROM {mock_session_model.__tablename__}"  # No WHERE clause
+    assert str(delete_stmt) == f"DELETE FROM {mock_session_model.__tablename__}"
     mock_sync_session.commit.assert_called_once()
 
 
@@ -726,7 +720,7 @@ def test_sync_backend_internal_delete_expired_sync(
     mock_sync_config: MagicMock,
     mock_session_model: type[SessionModelMixin],
     monkeypatch: MonkeyPatch,
-    mock_sync_session: MagicMock,  # Inject mock session
+    mock_sync_session: MagicMock,
 ) -> None:
     """Test the internal _delete_expired_sync method."""
     mock_sync_session = MagicMock(spec=SyncSession)
@@ -740,9 +734,462 @@ def test_sync_backend_internal_delete_expired_sync(
     sync_backend._delete_expired_sync()  # pyright: ignore [reportPrivateUsage]
 
     mock_sync_session.execute.assert_called_once()
-    # Check the delete statement structure loosely
     call_args, _ = mock_sync_session.execute.call_args
     delete_stmt = call_args[0]
     assert str(delete_stmt).startswith(f"DELETE FROM {mock_session_model.__tablename__}")
-    assert "WHERE" in str(delete_stmt) and "now() >" in str(delete_stmt)  # Match actual SQL
+    assert "WHERE" in str(delete_stmt) and "now() >" in str(delete_stmt)
     mock_sync_session.commit.assert_called_once()
+
+
+# --- SessionModelMixin Edge Cases ---
+
+
+class MockSessionModelEdgeCases:
+    """Test edge cases for SessionModelMixin."""
+
+    def test_table_args_with_spanner_dialect(self) -> None:
+        """Test table args generation for Spanner dialect."""
+        dialect_mock = Mock()
+        dialect_mock.name = "spanner+spanner"
+
+        result = MockSessionModel._create_unique_session_id_constraint(dialect=dialect_mock)
+        assert result is False
+
+        result = MockSessionModel._create_unique_session_id_index(dialect=dialect_mock)
+        assert result is True
+
+    def test_table_args_with_postgresql_dialect(self) -> None:
+        """Test table args generation for PostgreSQL dialect."""
+        dialect_mock = Mock()
+        dialect_mock.name = "postgresql"
+
+        result = MockSessionModel._create_unique_session_id_constraint(dialect=dialect_mock)
+        assert result is True
+
+        result = MockSessionModel._create_unique_session_id_index(dialect=dialect_mock)
+        assert result is False
+
+    def test_is_expired_expression(self) -> None:
+        """Test the SQL expression for is_expired."""
+        expr = MockSessionModel.is_expired
+        assert expr is not None
+        assert hasattr(expr, "expression")
+
+    def test_session_model_fields(self) -> None:
+        """Test that all required fields are present."""
+        now = datetime.datetime.now(datetime.timezone.utc)
+        session = MockSessionModel(
+            session_id="test_123",
+            data=b"test_data",
+            expires_at=now + datetime.timedelta(hours=1),
+        )
+
+        assert session.session_id == "test_123"
+        assert session.data == b"test_data"
+        assert session.expires_at > now
+        assert hasattr(session, "id")
+
+
+# --- Async Backend Error Tests ---
+
+
+class TestAsyncBackendErrors:
+    """Test error scenarios for async backend."""
+
+    @pytest.fixture()
+    def mock_async_config_errors(self) -> MagicMock:
+        """Create mock async config."""
+        config = MagicMock(spec=SQLAlchemyAsyncConfig)
+        session = AsyncMock(spec=AsyncSession)
+        session.__aenter__.return_value = session
+        session.__aexit__.return_value = None
+        config.get_session.return_value = session
+        return config
+
+    @pytest.fixture()
+    def async_backend_errors(self, mock_async_config_errors: MagicMock) -> SQLAlchemyAsyncSessionBackend:
+        """Create async backend with mock config."""
+        return SQLAlchemyAsyncSessionBackend(
+            config=ServerSideSessionConfig(max_age=3600),
+            alchemy_config=mock_async_config_errors,
+            model=MockSessionModel,
+        )
+
+    async def test_get_database_error(
+        self,
+        async_backend_errors: SQLAlchemyAsyncSessionBackend,
+        mock_async_config_errors: MagicMock,
+    ) -> None:
+        """Test handling of database errors during get operation."""
+        session = mock_async_config_errors.get_session.return_value
+
+        session.scalars.side_effect = Exception("Database connection lost")
+
+        with pytest.raises(Exception, match="Database connection lost"):
+            await async_backend_errors.get("session_123", Mock())
+
+    async def test_set_database_error_on_commit(
+        self,
+        async_backend_errors: SQLAlchemyAsyncSessionBackend,
+        mock_async_config_errors: MagicMock,
+    ) -> None:
+        """Test handling of database errors during set operation commit."""
+        session = mock_async_config_errors.get_session.return_value
+
+        mock_result = MagicMock()
+        mock_result.one_or_none.return_value = None
+
+        async def mock_scalars(*args: Any, **kwargs: Any) -> MagicMock:
+            return mock_result
+
+        session.scalars = mock_scalars
+        session.commit.side_effect = Exception("Commit failed")
+
+        with pytest.raises(Exception, match="Commit failed"):
+            await async_backend_errors.set("session_123", b"data", Mock())
+
+    async def test_delete_database_error(
+        self,
+        async_backend_errors: SQLAlchemyAsyncSessionBackend,
+        mock_async_config_errors: MagicMock,
+    ) -> None:
+        """Test handling of database errors during delete operation."""
+        session = mock_async_config_errors.get_session.return_value
+
+        session.execute.side_effect = Exception("Delete operation failed")
+
+        with pytest.raises(Exception, match="Delete operation failed"):
+            await async_backend_errors.delete("session_123", Mock())
+
+    async def test_concurrent_session_modification(
+        self,
+        async_backend_errors: SQLAlchemyAsyncSessionBackend,
+        mock_async_config_errors: MagicMock,
+    ) -> None:
+        """Test behavior with concurrent session modifications."""
+        session = mock_async_config_errors.get_session.return_value
+
+        mock_result1 = MagicMock()
+        mock_result1.one_or_none.return_value = MockSessionModel(
+            session_id="test",
+            data=b"data",
+            expires_at=datetime.datetime.now(datetime.timezone.utc) + datetime.timedelta(hours=1),
+        )
+
+        mock_result2 = MagicMock()
+        mock_result2.one_or_none.return_value = None
+
+        async def mock_scalars_1(*args: Any, **kwargs: Any) -> MagicMock:
+            return mock_result1
+
+        async def mock_scalars_2(*args: Any, **kwargs: Any) -> MagicMock:
+            return mock_result2
+
+        scalars_calls = iter([mock_scalars_1, mock_scalars_2])
+
+        async def mock_scalars_dispatcher(*args: Any, **kwargs: Any) -> MagicMock:
+            return await next(scalars_calls)(*args, **kwargs)
+
+        session.scalars = mock_scalars_dispatcher
+
+        result1 = await async_backend_errors.get("test", Mock())
+        assert result1 == b"data"
+
+        result2 = await async_backend_errors.get("test", Mock())
+        assert result2 is None
+
+
+# --- Sync Backend Error Tests ---
+
+
+class TestSyncBackendErrors:
+    """Test error scenarios for sync backend."""
+
+    @pytest.fixture()
+    def mock_sync_config_errors(self) -> MagicMock:
+        """Create mock sync config."""
+        config = MagicMock(spec=SQLAlchemySyncConfig)
+        session = MagicMock(spec=SyncSession)
+        session.__enter__.return_value = session
+        session.__exit__.return_value = None
+        config.get_session.return_value = session
+        return config
+
+    @pytest.fixture()
+    def sync_backend_errors(self, mock_sync_config_errors: MagicMock) -> SQLAlchemySyncSessionBackend:
+        """Create sync backend with mock config."""
+        return SQLAlchemySyncSessionBackend(
+            config=ServerSideSessionConfig(max_age=3600),
+            alchemy_config=mock_sync_config_errors,
+            model=MockSessionModel,
+        )
+
+    def test_get_sync_database_error(
+        self,
+        sync_backend_errors: SQLAlchemySyncSessionBackend,
+        mock_sync_config_errors: MagicMock,
+    ) -> None:
+        """Test handling of database errors during sync get operation."""
+        session = mock_sync_config_errors.get_session.return_value
+        session.scalars.side_effect = Exception("Database error")
+
+        with pytest.raises(Exception, match="Database error"):
+            sync_backend_errors._get_sync("session_123")
+
+    def test_set_sync_database_error(
+        self,
+        sync_backend_errors: SQLAlchemySyncSessionBackend,
+        mock_sync_config_errors: MagicMock,
+    ) -> None:
+        """Test handling of database errors during sync set operation."""
+        session = mock_sync_config_errors.get_session.return_value
+        mock_result = MagicMock()
+        mock_result.one_or_none.return_value = None
+        session.scalars.return_value = mock_result
+        session.commit.side_effect = Exception("Commit failed")
+
+        with pytest.raises(Exception, match="Commit failed"):
+            sync_backend_errors._set_sync("session_123", b"data")
+
+    def test_delete_sync_constraint_violation(
+        self,
+        sync_backend_errors: SQLAlchemySyncSessionBackend,
+        mock_sync_config_errors: MagicMock,
+    ) -> None:
+        """Test handling of constraint violations during delete."""
+        session = mock_sync_config_errors.get_session.return_value
+        session.execute.side_effect = Exception("Foreign key constraint violation")
+
+        with pytest.raises(Exception, match="Foreign key constraint violation"):
+            sync_backend_errors._delete_sync("session_123")
+
+
+# --- Configuration Edge Cases ---
+
+
+class TestConfigurationEdgeCases:
+    """Test configuration edge cases."""
+
+    def test_backend_with_minimal_max_age(self) -> None:
+        """Test backend behavior with minimal max_age."""
+        config = MagicMock(spec=SQLAlchemyAsyncConfig)
+        session = MagicMock(spec=AsyncSession)
+        session.__aenter__.return_value = session
+        session.__aexit__.return_value = None
+        config.get_session.return_value = session
+
+        backend = SQLAlchemyAsyncSessionBackend(
+            config=ServerSideSessionConfig(max_age=1),
+            alchemy_config=config,
+            model=MockSessionModel,
+        )
+
+        assert backend.config.max_age == 1
+
+    def test_backend_config_property_setter(self) -> None:
+        """Test that config property can be updated."""
+        config = MagicMock(spec=SQLAlchemyAsyncConfig)
+
+        backend = SQLAlchemyAsyncSessionBackend(
+            config=ServerSideSessionConfig(max_age=3600),
+            alchemy_config=config,
+            model=MockSessionModel,
+        )
+
+        new_config = ServerSideSessionConfig(max_age=7200)
+        backend.config = new_config
+        assert backend.config.max_age == 7200
+
+    def test_select_session_obj_query_generation(self) -> None:
+        """Test the SQL query generation for selecting session objects."""
+        config = MagicMock(spec=SQLAlchemyAsyncConfig)
+
+        backend = SQLAlchemyAsyncSessionBackend(
+            config=ServerSideSessionConfig(max_age=3600),
+            alchemy_config=config,
+            model=MockSessionModel,
+        )
+
+        query = backend._select_session_obj("test_session_id")
+        query_str = str(query)
+
+        assert "mock_session" in query_str
+        assert "session_id" in query_str
+        assert "WHERE" in query_str
+
+
+# --- Timezone Handling Tests ---
+
+
+class TestTimezoneHandling:
+    """Test timezone handling across Python versions."""
+
+    @pytest.mark.skipif(sys.version_info < (3, 11), reason="Testing Python 3.11+ behavior")
+    def test_get_utc_now_python_311_plus(self) -> None:
+        """Test get_utc_now for Python 3.11+."""
+        from advanced_alchemy.utils.time import get_utc_now
+
+        now = get_utc_now()
+        assert now.tzinfo is not None
+        assert now.tzinfo.utcoffset(None) == datetime.timedelta(0)
+
+    @pytest.mark.skipif(sys.version_info >= (3, 11), reason="Testing Python <3.11 behavior")
+    def test_get_utc_now_python_pre_311(self) -> None:
+        """Test get_utc_now for Python versions before 3.11."""
+        from advanced_alchemy.utils.time import get_utc_now
+
+        now = get_utc_now()
+        assert now.tzinfo is not None
+        assert now.tzinfo.utcoffset(None) == datetime.timedelta(0)
+
+    def test_session_expiry_timezone_aware(self) -> None:
+        """Test that session expiry is always timezone-aware."""
+        from advanced_alchemy.utils.time import get_utc_now
+
+        now = get_utc_now()
+        future = now + datetime.timedelta(hours=1)
+        past = now - datetime.timedelta(hours=1)
+
+        active_session = MockSessionModel(
+            session_id="active",
+            data=b"data",
+            expires_at=future,
+        )
+        expired_session = MockSessionModel(
+            session_id="expired",
+            data=b"data",
+            expires_at=past,
+        )
+
+        assert active_session.expires_at.tzinfo is not None
+        assert expired_session.expires_at.tzinfo is not None
+
+        assert not active_session.is_expired
+        assert expired_session.is_expired
+
+
+# --- Large Data Handling Tests ---
+
+
+class TestLargeDataHandling:
+    """Test handling of large session data."""
+
+    @pytest.fixture()
+    def large_data(self) -> bytes:
+        """Generate large data for testing."""
+        return b"x" * (1024 * 1024)
+
+    async def test_async_backend_large_data(
+        self,
+        large_data: bytes,
+    ) -> None:
+        """Test async backend with large data."""
+        config = MagicMock(spec=SQLAlchemyAsyncConfig)
+        session = MagicMock(spec=AsyncSession)
+        session.__aenter__.return_value = session
+        session.__aexit__.return_value = None
+        config.get_session.return_value = session
+
+        backend = SQLAlchemyAsyncSessionBackend(
+            config=ServerSideSessionConfig(max_age=3600),
+            alchemy_config=config,
+            model=MockSessionModel,
+        )
+
+        mock_result = MagicMock()
+        mock_result.one_or_none.return_value = None
+
+        async def mock_scalars(*args: Any, **kwargs: Any) -> MagicMock:
+            return mock_result
+
+        session.scalars = mock_scalars
+        session.commit = AsyncMock()
+
+        await backend.set("large_session", large_data, Mock())
+
+        session.add.assert_called_once()
+        added_obj = session.add.call_args[0][0]
+        assert added_obj.data == large_data
+
+    def test_sync_backend_large_data(
+        self,
+        large_data: bytes,
+    ) -> None:
+        """Test sync backend with large data."""
+        config = MagicMock(spec=SQLAlchemySyncConfig)
+        session = MagicMock(spec=SyncSession)
+        session.__enter__.return_value = session
+        session.__exit__.return_value = None
+        config.get_session.return_value = session
+
+        backend = SQLAlchemySyncSessionBackend(
+            config=ServerSideSessionConfig(max_age=3600),
+            alchemy_config=config,
+            model=MockSessionModel,
+        )
+
+        mock_result = MagicMock()
+        mock_result.one_or_none.return_value = None
+        session.scalars.return_value = mock_result
+
+        backend._set_sync("large_session", large_data)
+
+        session.add.assert_called_once()
+        added_obj = session.add.call_args[0][0]
+        assert added_obj.data == large_data
+
+
+# --- Session ID Validation Tests ---
+
+
+class TestSessionIDValidation:
+    """Test session ID handling edge cases."""
+
+    @pytest.mark.parametrize(
+        "session_id",
+        [
+            "",
+            "a" * 256,
+            "session with spaces",
+            "session/with/slashes",
+            "session?with=query",
+            "session#with#hash",
+            "😀emoji-session",
+            "\nsession\nwith\nnewlines",
+        ],
+    )
+    async def test_various_session_ids(
+        self,
+        session_id: str,
+    ) -> None:
+        """Test handling of various session ID formats."""
+        config = MagicMock(spec=SQLAlchemyAsyncConfig)
+        session = MagicMock(spec=AsyncSession)
+        session.__aenter__.return_value = session
+        session.__aexit__.return_value = None
+        config.get_session.return_value = session
+
+        backend = SQLAlchemyAsyncSessionBackend(
+            config=ServerSideSessionConfig(max_age=3600),
+            alchemy_config=config,
+            model=MockSessionModel,
+        )
+
+        mock_result = MagicMock()
+        mock_result.one_or_none.return_value = None
+
+        async def mock_scalars(*args: Any, **kwargs: Any) -> MagicMock:
+            return mock_result
+
+        session.scalars = mock_scalars
+        session.commit = AsyncMock()
+
+        await backend.set(session_id, b"data", Mock())
+
+        session.add.assert_called_once()
+        added_obj = session.add.call_args[0][0]
+
+        if len(session_id) > 255:
+            assert len(added_obj.session_id) <= 255
+        else:
+            assert added_obj.session_id == session_id
